# docs/swagger.yaml
openapi: 3.1.0
info:
  title: Commerce API
  description: "REST API para gestao de operacoes de comercio: catalogo, carrinho,\
    \ pedidos, pagamentos e relatorios."
  contact:
    name: Equipe Commerce API
    url: https://empresa.com/commerce-api
    email: commerce-api@empresa.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: "1.0"
...
servers:
- url: http://localhost:8080
  description: Ambiente local
- url: https://api-dev.empresa.com
  description: Ambiente de desenvolvimento
security:
- bearerAuth: []
...
paths:
  /users/{id}:
    get:
      tags:
      - Administração - Usuários
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      tags:
      - Administração - Usuários
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      tags:
      - Administração - Usuários
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /promotions/{id}:
    get:
      tags:
      - Público - Catálogo
      summary: Consultar promoção por ID
      operationId: findById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PromotionResponse"
    put:
      tags:
      - Administração - Catálogo
      summary: Atualizar promoção
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromotionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PromotionResponse"
    delete:
      tags:
      - Administração - Catálogo
      summary: Excluir promoção
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /products/{id}:
    get:
      tags:
      - Público - Catálogo
      summary: Consultar produto por ID
      operationId: findById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
    put:
      tags:
      - Administração - Catálogo
      summary: Atualizar produto
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
    delete:
      tags:
      - Administração - Catálogo
      summary: Excluir produto
      operationId: delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /expenses/{id}:
    get:
      tags:
      - Administração - Despesas
      operationId: findById_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExpenseResponse"
    put:
      tags:
      - Administração - Despesas
      operationId: update_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExpenseResponse"
    delete:
      tags:
      - Administração - Despesas
      operationId: delete_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /categories/{id}:
    get:
      tags:
      - Público - Catálogo
      summary: Consultar categoria por ID
      operationId: findById_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
    put:
      tags:
      - Administração - Catálogo
      summary: Atualizar categoria
      operationId: update_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
    delete:
      tags:
      - Administração - Catálogo
      summary: Excluir categoria
      operationId: delete_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /brands/{id}:
    get:
      tags:
      - Público - Catálogo
      summary: Consultar marca por ID
      operationId: findById_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BrandResponse"
    put:
      tags:
      - Administração - Catálogo
      summary: Atualizar marca
      operationId: update_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BrandResponse"
    delete:
      tags:
      - Administração - Catálogo
      summary: Excluir marca
      operationId: delete_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /reports/sales:
    post:
      tags:
      - Administração - Relatórios de Vendas
      operationId: summary
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesReportRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SalesSummaryResponse"
...
  /promotions:
    get:
      tags:
      - Público - Catálogo
      summary: Listar promoções
      operationId: findAll_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromotionResponse"
    post:
      tags:
      - Administração - Catálogo
      summary: Cadastrar promoção
      operationId: save_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromotionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PromotionResponse"
  /promotions/link-product:
    post:
      tags:
      - Administração - Catálogo
      summary: Vincular promoção a produto
      operationId: linkToProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkPromotionToProductRequest"
        required: true
      responses:
        "200":
          description: OK
  /products:
    get:
      tags:
      - Público - Catálogo
      summary: Listar produtos
      operationId: findAll_2
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductResponse"
    post:
      tags:
      - Administração - Catálogo
      summary: Cadastrar produto
      operationId: save_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductResponse"
...
  /financial/summary:
    post:
      tags:
      - Administração - Financeiro
      operationId: summary_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialSummaryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FinancialSummaryResponse"
  /expenses:
    get:
      tags:
      - Administração - Despesas
      operationId: findAll_3
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseResponse"
    post:
      tags:
      - Administração - Despesas
      operationId: save_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExpenseResponse"
  /categories:
    get:
      tags:
      - Público - Catálogo
      summary: Listar categorias
      operationId: findAll_4
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
    post:
      tags:
      - Administração - Catálogo
      summary: Cadastrar categoria
      operationId: save_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
  /brands:
    get:
      tags:
      - Público - Catálogo
      summary: Listar marcas
      operationId: findAll_5
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrandResponse"
    post:
      tags:
      - Administração - Catálogo
      summary: Cadastrar marca
      operationId: save_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBrandRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BrandResponse"
...
  /orders/{orderId}:
    get:
      tags:
      - Administração - Pedidos
      operationId: findById_6
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponse"
...
components:
  schemas:
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        cpf:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - cpf
      - email
      - name
      - password
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        cpf:
          type: string
        email:
          type: string
        active:
          type: boolean
        role:
          type: string
          enum:
          - ROLE_SUPER_ADMIN
          - ROLE_ADMIN
          - ROLE_VIEWER
          - ROLE_CUSTOMER
    CreatePromotionRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        discountType:
          type: string
          enum:
          - PERCENTAGE
          - FIXED_VALUE
        discountValue:
          type: number
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
    PromotionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        discountType:
          type: string
          enum:
          - PERCENTAGE
          - FIXED_VALUE
        discountValue:
          type: number
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
    CreateProductAttributeRequest:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        price:
          type: number
        stock:
          type: integer
          format: int32
        active:
          type: boolean
        categoryId:
          type: integer
          format: int64
        brandId:
          type: integer
          format: int64
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/CreateProductAttributeRequest"
      required:
      - active
      - brandId
      - categoryId
      - description
      - name
      - price
      - stock
    ProductAttributeResponse:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
          format: int32
        active:
          type: boolean
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        brandId:
          type: integer
          format: int64
        brandName:
          type: string
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/ProductAttributeResponse"
        finalPrice:
          type: number
        promotionActive:
          type: boolean
...
    CreateExpenseRequest:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        category:
          type: string
          enum:
          - COMPRA_PRODUTOS
          - MARKETING
          - EMBALAGENS
          - INFRAESTRUTURA
          - FRETE
          - IMPOSTOS
          - OUTROS
        expenseDate:
          type: string
          format: date
    ExpenseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        amount:
          type: number
        category:
          type: string
          enum:
          - COMPRA_PRODUTOS
          - MARKETING
          - EMBALAGENS
          - INFRAESTRUTURA
          - FRETE
          - IMPOSTOS
          - OUTROS
        expenseDate:
          type: string
          format: date
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        active:
          type: boolean
      required:
      - active
      - name
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        active:
          type: boolean
    CreateBrandRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        active:
          type: boolean
      required:
      - active
      - name
    BrandResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        active:
          type: boolean
    SalesReportRequest:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        saleChannel:
          type: string
          enum:
          - ONLINE
          - MANUAL
    SalesSummaryResponse:
      type: object
      properties:
        totalSales:
          type: integer
          format: int64
        onlineSales:
          type: integer
          format: int64
        manualSales:
          type: integer
          format: int64
        totalRevenue:
          type: number
        onlineRevenue:
          type: number
        manualRevenue:
          type: number
        averageTicket:
          type: number
...
    LinkPromotionToProductRequest:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        promotionId:
          type: integer
          format: int64
...
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productName:
          type: string
        productPrice:
          type: number
        quantity:
          type: integer
          format: int32
        subtotal:
          type: number
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
        addressId:
          type: integer
          format: int64
        total:
          type: number
        status:
          type: string
          enum:
          - PENDENTE
          - PAGO
          - ENVIADO
          - ENTREGUE
          - CANCELADO
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
...
    FinancialSummaryRequest:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        saleChannel:
          type: string
          enum:
          - ONLINE
          - MANUAL
        expenseCategory:
          type: string
          enum:
          - COMPRA_PRODUTOS
          - MARKETING
          - EMBALAGENS
          - INFRAESTRUTURA
          - FRETE
          - IMPOSTOS
          - OUTROS
        includeAffiliateCommissions:
          type: boolean
    FinancialSummaryResponse:
      type: object
      properties:
        totalRevenue:
          type: number
        onlineRevenue:
          type: number
        manualRevenue:
          type: number
        totalExpenses:
          type: number
        affiliateCommissions:
          type: number
        resultadoBruto:
          type: number
        resultadoLiquido:
          type: number
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT